added mobile animations

+++ scripts/login.js
@@ -1,7 +1,7 @@
 //Globals
 let users = [];
 let loginForm = document.getElementById("loginForm");
-
+let loginBox = document.querySelector(".login-box");
 function templateForgotPassword() {
   return `
         <img onclick="showLogin();" src="./img/arrow-left-line.svg" class="arrow-back" />
@@ -9,7 +9,7 @@ function templateForgotPassword() {
         <span class="login-line"></span>
         <p class="mt-5">Don't worry! We will send you an email with the instructions to reset your password.</p>
     
-        <form onsubmit="return false;">
+        <form onsubmit="sendPasswordMail(event);">
             <div class="input-container reset-password-input">
                 <input
                 class="text-input"
@@ -17,6 +17,7 @@ function templateForgotPassword() {
                 name="email"
                 id="email"
                 placeholder="Email"
+                required
                 />
                 <img
                 src="./img/mail.svg"
@@ -24,63 +25,71 @@ function templateForgotPassword() {
                 class="input-icons"
                 />
             </div>
-            <button class="signup-btn reset-password-btn">Send me the email</button>
+            <button type="submit" class="signup-btn reset-password-btn">Send me the email</button>
         </form>
     `;
 }
 
 function templateLogin() {
   return `
-    <form onsubmit="return false;" id="loginForm">
-          <h1 class="">Log in</h1>
-          <span class="login-line"></span>
-          <div class="input-container">
-            <input
-              class="text-input"
-              type="email"
-              name="email"
-              id="email"
-              placeholder="Email"
-            />
-            <img
-              src="./img/mail.svg"
-              alt="Small Mail Symbol"
-              class="input-icons"
-            />
-          </div>
-          <div class="input-container">
-            <input
-              class="text-input"
-              type="password"
-              name="password"
-              placeholder="Password"
-              id="password"
-            />
-            <img
-              src="./img/lock.svg"
-              alt="Small Lock Symbol"
-              class="input-icons"
-            />
-          </div>
-          <div class="checkbox-container">
-            <div class="remember-me-container">
-              <input
-                class="remember-checkbox"
-                type="checkbox"
-                name="remember"
-                id="remember"
-              />
-              <span>Remember me</span>
-            </div>
-            <a onclick="forgotMyPassword();" class="forgot-link"
-              >Forgot my password</a
-            >
-          </div>
-          <div class="login-btn-container">
-            <button class="signup-btn">Log in</button>
-            <button class="signup-btn guest-login-btn">Guest Log in</button>
-          </div>
-        </form>
+  <form onsubmit="login(); return false;" id="loginForm">
+  <h1 class="">Log in</h1>
+  <span class="login-line"></span>
+  <div class="input-container">
+    <input
+      class="text-input"
+      type="email"
+      name="email"
+      id="email"
+      placeholder="Email"
+      required
+    />
+    <img
+      src="./img/mail.svg"
+      alt="Small Mail Symbol"
+      class="input-icons"
+    />
+  </div>
+  <div class="input-container">
+    <input
+      class="text-input"
+      type="password"
+      name="password"
+      placeholder="Password"
+      id="password"
+      required
+    />
+    <img
+      src="./img/lock.svg"
+      alt="Small Lock Symbol"
+      class="input-icons"
+    />
+  </div>
+  <div class="checkbox-container">
+    <div class="remember-me-container">
+      <input
+        class="remember-checkbox"
+        type="checkbox"
+        name="remember"
+        id="remember"
+      />
+      <span>Remember me</span>
+    </div>
+    <a onclick="forgotMyPassword();" class="forgot-link"
+      >Forgot my password</a
+    >
+  </div>
+  <div class="login-btn-container">
+    <button class="signup-btn">Log in</button>
+    <button
+      class="signup-btn guest-login-btn"
+      type="button"
+      onclick="guestLogin();"
+    >
+      Guest Log in
+    </button>
+  </div>
+</form>
     `;
 }
 
@@ -161,22 +170,24 @@ function templateSignUp() {
 }
 
 function forgotMyPassword() {
-  loginForm.innerHTML = "";
-  loginForm.innerHTML = templateForgotPassword();
+  let loginBox = document.querySelector(".login-box");
+  loginBox.innerHTML = "";
+  loginBox.innerHTML = templateForgotPassword();
 }
 
 function showLogin() {
-  loginForm.innerHTML = "";
-  loginForm.innerHTML = templateLogin();
+  loginBox.innerHTML = "";
+  loginBox.innerHTML = templateLogin();
 }
 
 function showSignup() {
-  loginForm.innerHTML = "";
-  loginForm.innerHTML = templateSignUp();
+  loginBox.innerHTML = "";
+  loginBox.innerHTML = templateSignUp();
 }
 
 async function init() {
   loadUsers();
+  checkIfUserIsLoggedIn();
 }
 
 async function loadUsers() {
@@ -191,7 +202,6 @@ function confirmPasswordIsSame(password, confirmedPassword) {
   if (password == confirmedPassword) {
     return true;
   } else {
-    alert("Your passwords are not similar!");
     return false;
   }
 }
@@ -203,14 +213,24 @@ async function registerUser() {
   let password = document.getElementById("password");
   let confirmPassword = document.getElementById("confirmPassword");
 
-  users.push({
-    name: name.value,
-    email: email.value,
-    password: password.value,
-  });
+  if (checkIfUserExists(email.value)) {
+    showLogin();
+    showTopDown("This User exists, please log in!");
+    signupBtn.disabled = false;
+    return 0;
+  }
 
   if (confirmPasswordIsSame(password.value, confirmPassword.value)) {
+    users.push({
+      name: name.value,
+      email: email.value,
+      password: password.value,
+    });
     await setItem("users", JSON.stringify(users));
+  } else {
+    showTopDown("Your passwords are not similar!");
+    signupBtn.disabled = false;
+    return 0;
   }
 
   resetForm(name);
@@ -219,21 +239,67 @@ async function registerUser() {
   resetForm(confirmPassword);
   signupBtn.disabled = false;
   showLogin();
+  showTopDown("You are registered!");
+}
+
+function checkIfUserExists(email) {
+  return users.some((user) => user.email === email);
 }
 
 function resetForm(input) {
   input.value = "";
 }
 
+function getLoginFormInput() {
+  const input = {
+    email: document.getElementById("email").value,
+    password: document.getElementById("password").value,
+    remember: document.getElementById("remember").checked,
+  };
+  return input;
+}
+
 function login() {
-  let email = document.getElementById("email");
-  let password = document.getElementById("password");
+  let formInput = getLoginFormInput();
+  let user = users.find((user) => user.email === formInput.email);
 
-  let user = users.find((user) => user.email === email.value);
-
-  if (user && user.password === password.value) {
+  if (user && user.password === formInput.password) {
     alert("You are logged in!");
+    if (formInput.remember) {
+      saveRememberMe(user);
+    }
   } else {
     alert("Your email or password is wrong!");
   }
+}
+
+function saveRememberMe(user) {
+  localStorage.setItem("currentUser", JSON.stringify(user));
+}
+
+function checkIfUserIsLoggedIn() {
+  let user = JSON.parse(localStorage.getItem("currentUser"));
+  if (user === null) {
+    console.info("No User was saved in local storage");
+  } else {
+    // window.location.href = "summary.html";
+  }
+}
+
+function showTopDown(message) {
+  let popup = document.getElementById("topdownMessages");
+  popup.innerHTML = message;
+  popup.classList.add("show-topdown");
+  setTimeout(() => {
+    popup.classList.remove("show-topdown");
+  }, 5000);
+}
+
+function guestLogin() {
+  window.location.href = `summary.html?name='Guest'`;
+}
+
+function sendPasswordMail(e) {
+  e.preventDefault();
+  showTopDown("We send you an email!");
 }
